Initial builder implementation
Implements a `Builder` that tracks what fields have been filled in so
far. Uses `Proxy` to provide a handler that can create methods to fill
in the fields when called as methods on the builder. Allows filling in
default values for a subset of parameters using `builderDef`.

One downside to the current setup is that when you're missing required
fields, the compiler errors are very long, and only the last line is
relevant.

Another downside is that `builderDef` can infer the type of the default
arguments, but it's difficult to require the first type parameter while
inferring the second one. Currently, the type of the defaults must be
explicitly specified using `Pick` or equivalent.