Default builder fields to the type being constructed
A `Builder<C, T>` should default to `Builder<C, T, {}>`, and a
`Builder<C>` should default to `Builder<C, C, {}>`; if the type being
built is the same as the one constructed, you shouldn't have to specify
it twice in the `Builder` signature.

Another alternative is to move `C` somewhere else in the declaration and
default `C = T`, since this would discourage users from putting a class
in the fields parameter (requiring building methods, etc.), but I think
this has the same effect. After all, nothing's stopping someone from
specifying a class type in the current `T` field, either.